requirements.txt
streamlit
python-docx
PyPDF2